Login backend, refactoring and fixes related to the router warnings
The backend for #27 is now ready for the session ID implementation(when I finish it). As for the warnings in #42 , they're now fixed. Removing "ReactDOM.render()" and only using render.root() fixed most warnings. The last warning that remained was fixed by importing createRoot from "react-dom/client" instead of "react-dom". Also, the "invalid email or password" alert now appears when the email and/or the password are incorrect. Furthermore, I added comments and refactored some of the code.